requirements.txt
Django==3.0.5
Pillow==7.1.2
djangorestframework==3.11.0
django-cors-headers==3.2.1